Added bucket-level preferred port mechanism to RESTful interface.  Closed questions.
After a discussion between Steve Yen and I, we decided the type of conneciton a 
smarter client may make may change over the course of time.  Even if it doesn't, we
may want this to be configuration level work, not code.  Thus, adding some configuration
at the bucket level is useful.