Modified JSON names to camelCase; new restrictions for bucket/pool names.
Since most consumers of the JSON interface are expected to be Java and C# 
users over the long haul, I made all multiWord JSON object names CamelCase.
This seems to be the most common convention.  The exceptions seem to be
when the RESTful interface is written in a language where _ or - is more
common.  However, it felt more like it should be consistent with what consumers
want, not server implementors.