Added hack to signal the end of a Comet response.
I've sent off a note to the mochiweb group, but the common thing to do
with Comet streaming is to send a "zero chunk" at the end.  Mochiweb
seems to lack a way to do this though.

For now, a response with four newlines indicates the end of a
JSON response.