downstream mbox monitoring window fixed
Previously, there was a window between monitoring a MBox
and sending a message to the MBox where the downstream
connection would close, and the caller would get an
entirely new MBox between monitoring and sending.

Now, a snapshot of the MBoxPid is taken and used consistently
for both monitoring and for sending, so there's no window
for race conditions.