Separated configuration replication into its own module.
While a simple change, this had a ripple effect:

* Configuration replication is now done using a supervised process.

If the configuration replicator fails for any reason, the supervisor
will restart it.

* Configuration replication is always scheduled.

A flapping node or unexpected messages to ns_node_disco would've
prevented periodic config synchronization.  Being done in its own
process with no other dependencies means this is guaranteed to occur.

* Emphemeral processes are no longer needed for config replication.

The server is long-running and exposes its API as server casts so all
config replication is serialized and asynchronous and occurs from
within the same process.

* ns_server_init is no longer needed

It was a transient supervised child that only existed a sort of
last-to-run hook during startup that would be run for its side-effects.

Instead, the primary responsbility of this (to replicate and announce
config) is now the responsibility of the initialization of
ns_config_rep.