Added cache reserved and allocated to UI and service.
This feature allows a user to see, from the status on the web UI or
from the REST service cluster cache reserved and the cluster cache
allocated.  Cache reserved is the sum total of all buckets on all
nodes.  Cache allocated is the amount of the cache identified as used
by underlying memcached stats calls on the bucket instances.