Don't specify lookup (reverts 0a2ffb83580b)
This cost me a lot of time trying to figure out why I couldn't get
resolution working.  I'm suspecting it has something to do with the
trouble a few other people were having as well.

It *may* be appropriate to change this in a few installations, but
disabling the default (native) by default is likely to break things.

BTW, this is how you see what the current value is:

     inet_db:res_option(lookup)

native is probably a good thing to add in there even if we do decide
to do a full list.