Fixed memory utilization to agree between graphs.
I didn't spend a lot of time investigating why, but the upper
graph seemed to be correct while the lower one was not.  I just
copied the logic straight from the upper one to the lower.

The upper one is a summary of all nodes, however.  The lower is
the per-node memory usage.