Don't send bogus stats for buckets that aren't in memcached
mc_client_binary always returns "ok" unless there was some network
error. The app is apparently responsible for looking at the status code
in the header if it's interested in the return value.

Since stats_collector was just looking for the ok in the return from
select_bucket, if the bucket didn't exist it just stayed in the previous
bucket and sent the aggregator the stats for that bucket. When the bytes
got added together, it ended up adding up the bytes for the default
bucket multiple times.

I changed stats_collector to just punt if the bucket doesn't exist. If a
bucket disappears from memcached but is still configured for some
reason, the stats will hang around in stats_aggregator, however.

Fixes bug 431