Log to files in ./logs when using start_shell.sh
This is primarily to test binary file-based logging operationally. I've
also created the shell scripts dump_logs.sh that just writes the logs in
the more familiar text form to stdout, and grep_logs.sh that greps for a
particular regular expression in the log entries, printing the entire
log entry instead of just the matching line.