Force memcached's bucket config to be what we want.
I do this in the stats collector, which is not ideal, but the stats
collector already has all the information we need, so it's convenient.
It also is already hitting every server every second and runs in only
one place.

If we see the size change (should be rare) it deletes and recreates the
bucket. This even works with the default bucket, obviating the need for
an environment variable (which only does anything with auto_create
anyway).

It won't delete the last bucket due to the way it gets the list of
servers, but as soon as a new bucket is created, it'll delete the last
one that was deleted, so the end results will be correct.

I took the cache_size stuff out of the isasl.pw since it only has any
effect with autocreate. Also updated the tests.