Change join/leave strategy
ns_cluster is now a gen_server that supervises ns_server_sup. To join a
cluster, we change the cookie to the new cookie in ns_config and erlang,
then instruct ns_server_sup to exit. To leave, we change the cookie,
disconnect all the nodes, instruct ns_server_sup to exit.

The API is the same, though unexpected things may happen since the
behavior is now different.