Send shutdown exit to ns_server_sup from ns_cluster.
When a supervisor gets an exit message from a child, it restarts that
child, unless the reason is shutdown. But since we can't pass any data
in the shutdown reason, that means ns_cluster needs to store state about
what it's doing anyway, so we have the exit message get sent from
ns_cluster and update the state accordingly.