Trap exits from the loop_out processes so we see normal exits.
By default, any exit type other than normal propagates between linked
processes.

The only way to get normal exits is to trap_exit which requires
careful manual processing of exit messages.

For the purposes of mc_server_*, this needs to happen in just a few
places, so it should be fairly safe.