Turn auto_create back on
We have a race where if a bucket is created and something tries to
connect to the bucket between the time the isasl.pw is written out/read
by memcached and the time the stats collector gets around to creating
the bucket, we end up with an mc_downstream child with no bucket.

We still have a situation where if you delete a bucket after talking to
it and then recreate it, there will be an mc_downstream hanging around
talking to the old bucket. It might actually be best to handle this in
memcached itself, disconnecting any clients still attached to a deleted
bucket.