Excise the NaNs from Web UI.  Bug 593.
While there's still lack of clarity for memory status for a down
node (for instance, it may show as memory available with a down
notification, if memcached crashed), the decision was made to just
excise the NaNs from the response.

In testing I found one other case which would show Inifinity.  I've
normalized both to 0.

This was tested with FireFox 3.5 and IE 6.