Implemented very basic diagnostics tool.
Signed-off-by: Matt Ingenthron <ingenthr@cep.net>

Note, this is related to enhancement in bug 694.  Normally this wouldn't
be cherry-picked over until completion, but the value of this in assisting
our debugging efforts makes it worthwhile.  The remainder of the
implementation will come over in a different commit and  then the bug
number will show in the summary.
  Matt Ingenthron <ingenthr@cep.net>