[cells] invoke callbacks in 'quiescent' state. References bug 706.
previously due to cell value propagation delay it was possible to be
running callback for cell which value if calculated now would be
undefined.

It was possible to trigger (harmless, it seems) exceptions due to
this. For example, queued update of analytics graph would run even if
user already switched UI to another section.

Signed-off-by: Matt Ingenthron <ingenthr@cep.net>