Bug 843: deduplicate ns_log messages
Currently we garbage collect the duplicates once a minute, generating a
log message with the number of repetitions for anything repeated more
than once. Nothing is logged more than once per five minute interval.
These are right at the top of the file so they can be changed easily,
though it's important to note that the GC interval is in milliseconds
while the time over which we suppress duplicates is in microseconds.

Signed-off-by: Dustin Sallings <dustin@spy.net>