Reimplemented ns_cluster:leave and took logic out of menelaus - bug 864
Leave was a really complicated interaction that involved telling
remote nodes to tell the local node to stop participating, but it left
the configs on both sides in bad states.

Now leave selects a remote node at random and has it do a config
update and sync and then immediately resets the local config to a
minimal state.

Signed-off-by: Steve Yen <steve.yen@gmail.com>