Bug 898: set initial value properly on receiving stats for a new bucket
From the bug:

I made a mistake eliminating the creation of the default ringdict
for every call to received; the "initial" argument to dict:update
is actually set in the dictionary, *not*  passed to the function,
so it was setting "undefined" in the values for the first sample.
The next sample would fix things, but if you had the stats screen
up, it would never heal.

Signed-off-by: Steve Yen <steve.yen@gmail.com>