Bug 907: merge stats for the same top key on multiple nodes.
This turned out to be harder than I thought; we had to go back
to using dicts and add each of the stats independently, and switch from
computing the ratio in the aggregator to storing both hits and misses so
we could add them independently and get an overall ratio. Also needed
changes to menelaus_stats to handle getting a dictionary and getting
hits and misses instead of the ratio from the aggregator.

Resubmitting because it didn't get cherry-picked properly.

Change-Id: I000bbe602577736595e3fd4287fef743fba1031b
Reviewed-on: https://review.northscale.com:8443/1227
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>