bug 912, crash in stats_collector, logging with wrong # of args
There's was a minor logging bug/crash in stats_collector, in the
message flush()'ing code, since there were the wrong number of
arguments to the logger.

The theory on what happened next: the badarg crash then crashes the
stats_collector gen_server.  Because of that, the stats_collector
doesn't get around to creating/synchronizing buckets in the memcached
process, and then things don't work so well after that.

A post-1.0.2 change would include moving the memcached bucket
synchronization code out of stats_collector.

Change-Id: I4602caf25b1e8be99aac5dd10f629590d5d12ce8
Reviewed-on: https://review.northscale.com:8443/1203
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>