rework of erlang name assignment logic (bugs 757, 995 and others)
This patch changes the way ns_server assigns erlang name. When clean
and fresh node starts it assigns self ns_1@127.0.0.1 name. When this
node joins existing cluster or some other nodes joins this node we
re-assign name. We connect to erlang port mapper on other node and get
source address of that connection. We use this source address to build
node name. We don't change node name after that anymore.

As a result of that dist_manager is not application anymore but a
single gen_server taking care of node name.

This approach allows us to pick meaningful and correct ip address. So
that both nodes ensure that they can connect to each other. See bug
995 for explanation of why it's important.

Change-Id: I7f9ea7b8d0f79b1baad7f173d932159c4d3e3397
Reviewed-on: http://review.northscale.com:8080/37
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Sean Lynch <seanl@literati.org>