bucketsStreamingConfig REST api
Allows ns_server to respond to a new REST API call of
/pools/[pool-name]/bucketsStreamingConfig/[bucket-name]

For example...
/pools/default/bucketsStreamingConfig/default

In contrast to the existing /pools/default/bucketsStreaming/default,
the response to the new REST call provides a subset of information.
The objective is the response value to bucketsStreamingConfig should
be stable between calls if the configuration has not changed.  This
allows REST clients, for example, to do a simple strcmp() to detect
whether configuration has actually changed.

See http://hub.northscale.com/wiki/index.php?title=MembaseOneRest
for more information.

Change-Id: I050fcf102a19bc9784e153ed47fede4e0c1c8bb1
Reviewed-on: http://review.northscale.com:8080/136
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>