saving initStatus state in initial config wizard screens
The objective is to allow the user to return to the same
initial configuration wizard dialog page, even if their
web browser restarts.

Implemented /node/controller/initStatus POST,
and initStatus field in /pools GET.

On the server-side, init_status value is saved in ns_config.

Implemented Next and Back buttons navigation (while actual
validation still on the Next clicks still to be done).

Change-Id: I4d9290fa993bdcec48164dcfe4cf60d901f72b18
Reviewed-on: http://review.northscale.com:8080/164
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>