license screen does REST calls to placeholder erlang modules
Implemented REST calls to /node/controller/license to do license
validation on the server-side (which is just a placeholder in erlang
for now).

Also, started moving away from DOM id's towards classes,
to allow for reusuability of the same rendered dialog pages in
different contexts: the initial config wizard and the
settings pages.

Change-Id: Ia9d345e8c01c8a057ad1f1409890d3f76a15382e
Reviewed-on: http://review.northscale.com:8080/173
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>