Bug 1256 - server not starting due to wrong memcached, moxi, engine paths
Recently, the linux wallace installer was changed to write out just an
empty priv/config file, in the hopes of reducing all the copies of
priv/config that we maintain.

However, wallace installs memcached, the engines, moxi, etc in certain
./bin/subdirectory places, and it's easier to fix bug 1256 by changing
the relative paths in ns_config_default in the port_servers config to
match wallace, rather than changing wallace.

Change-Id: I6227b284914f8838f95974458208b5d4cf0753fe
Reviewed-on: http://review.northscale.com:8080/399
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Sean Lynch <seanl@literati.org>