Make rebalance actually rebalance.
We also now map all the vbuckets even in a cluster of one, so the node
no longer shows up in "pending rebalance" state. The janitor also runs
every 10 seconds. It's fast, so I just run it synchronously. This
means if memcached restarts, the vbuckets will get set to the right
state within 10 seconds and all the migrators will be restarted
(provided they don't hang - this may be a bug).

Change-Id: I2cf353a22d1a839a1f1fc86298470379b3577e47
Reviewed-on: http://review.northscale.com:8080/648
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>