Only delete vbuckets that are actually there.
Previously we'd always set vbuckets to dead, then delete them, then
set them to replica state. This seems to occasionally result in a race
condition in the EP engine where we end up with the vbucket staying in
dead state. We'd clean it up, but with lots of data it could keep
doing that.

Change-Id: I39e39b44ef3594a5a2abaf57522e7c96642d47fc
Reviewed-on: http://review.northscale.com:8080/753
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>