Bug 1485: Fix abuse of shun to eject nodes we can talk to
ns_cluster:shun was only ever intended to eject nodes we couldn't talk
to, and it's called by leave from the leaving node when it's leaving
to remove it from nodes_wanted. It does *nothing* to the node that's
being shunned, leaving it with an old config. Added leave(Node) to
force a node to leave.

Change-Id: Iae4dd0eca752c8a2e81dd1542c498edc12693a4c
Reviewed-on: http://review.northscale.com:8080/754
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>