Bug 1617: Merge on config pull instead of replacing
In making set use update instead of merge, I missed that ns_config_rep
was using set with a kvlist in order to merge the config from the
other node in the pull case (push case was fine because I had not
updated set_remote).

I added a merge function and renamed the still-working set_remote to
merge_remote to better reflect what it's actually doing.

This is another bug that would have been causing vclocks to increment
into the stratosphere.

Change-Id: Idcd7347a3086cf34ff83489c8cfff8b5fe6ee499
Reviewed-on: http://review.northscale.com/940
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>