Distributed, replicated logs.
Instead of having ns_log run in the global singleton, ns_log,
ns_log_events, ns_mail_sup, and ns_mail now run on each node. Log
messages are deduplicated locally and broadcast to the entire cluster,
and each log server gossips with the others periodically to make sure
it didn't miss any.

This is for beta2, not for beta1.1, so hold off on approving it until
after beta 1.1 goes out.

Change-Id: I1fd2e6a8145b21cf608dcd4ba1f35f1574a7b0e2
Reviewed-on: http://review.northscale.com/945
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>