Bug 1634: Fix badmatch in ns_config.
Apparently my assumptions about when we'd see two vclocks that descend
from one another are incorrect: we're seeing vclocks with the same
counters but different timestamps for some reason on the same
keys. Rather than crashing in this scenario, I log it, but only if the
actual config values are different.

Change-Id: I4ee795093efb5f8c34fda003c3538c9f91292fd1
Reviewed-on: http://review.northscale.com/958
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>