Bug 1658: Kill the global singleton supervisor if not registered
Somehow we ended up with a cluster where the global singleton
supervisor was running but not registered, but some of its children
were registered, preventing the global singleton supervisor from
coming up anywhere else in the cluster.

We should rethink the whole global singleton strategy, but not for
beta1.1.

Another symptom was that nodes that were trying to bring up the global
singleton unsuccessfully eventually gave up and ended up in a
brain-dead state. I've added a sleep to ns_cluster to hopefully
prevent this from happening.

The global singleton supervisor now spawns a watchdog process that
kills it if it finds the name unregistered.

Change-Id: I1e51daa3f44bdc44bb22c46442fb5d7327b5f825
Reviewed-on: http://review.northscale.com/987
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>