Bug 1662: Prevent ping-ponging configs
Under some circumstances, we could end up with vector clocks that were
in different orders but compared as descending from one another in
vclock.erl. Since merging vclocks doesn't put them in a canonical
order and since we weren't incrementing on top of merge, this would
cause the configs to ping pong back and forth.

We now always put the clocks in canonical order *and* always increment
on merge if the two versions don't compare equal. It's not really
necessary to canonicalize the order if we increment, but it's also not
harmful.

Change-Id: I1ec7688086fd21151018892fd17816633ad9f5f2
Reviewed-on: http://review.northscale.com/991
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>