Bug 624: persist ns_log entries
Now that there is a log server on each node with broadcasted/synced
logs, I've just made ns_log load from a configurable disk file on
startup and schedule a save 30 seconds after any change to the logs,
so we don't write multiple times if there is a burst of changes.

The file is configured in the ns_log/filename key/subkey in the
config. I've updated gen_cluster_scripts.rb to put in a per-node item
for this. Presumably Wallace will need to be updated as well.

Change-Id: I58e243790e0cf6a36a204c6bc01e4866c689373b
Reviewed-on: http://review.northscale.com/1002
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>