Bug 1682: Make ns_node_disco do more stuff asynchronously
From looking at the logs on bug 1682, I see lots of stuff timing out
trying to talk to ns_node_disco when a node is down, presumably
because ns_node_disco is busy waiting for ping responses from the
node. In addition, ns_node_disco was failing to start because its
first attempt to ping other nodes was timing out.

I moved the pings to a child process since they don't affect state,
and I made timeouts waiting for the initial call to
do_nodes_wanted_updated not result in a crash.

Change-Id: Iacb4246ae8449e7f0d6f96a606bcbb719cf30986
Reviewed-on: http://review.northscale.com/1014
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>