Add stats_collector and ns_stats.hrl
The include file has the record format for stats. Stats_collector is
the good ol' stats collector module, new and improved for the second
decade of the twenty-first century.

There should now be a stats collector per node per bucket, and once a
second the stats collector will get the stats for that node for that
bucket and generate an ns_stats_event event with the stats record.

Change-Id: Ide42368f424e94f31ca4f203e4653263c86a3fd6
Reviewed-on: http://review.northscale.com/1035
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>