Add centralized tick server.
This includes an event manager, ns_tick_event, on each node, and a
central gen_server, ns_tick, that generates an event {tick, TS} on
every node every second, where TS is the output of erlang:now() on the
node generating the tick. This means the timestamps can move around if
the time changes on the node where the gen_server is running, or if a
crash causes the gen_server to move.

To get the current time according to the tick service, call
ns_tick:time().

Currently the envisioned subscribers to the tick service will be
robust to changes in the time.

Change-Id: I7d856ed4aab654489d3799bd46f7bd0df71c450d
Reviewed-on: http://review.northscale.com/1062
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>