Bug 1474: Real progress bars
This presently has a bug where sometimes certain nodes' progress bars
will be seen to go from 0% to 100% twice if we're evacuating nodes
along with any other rebalance operation because I count the number of
remaining moves as moves both from and to a node. This may not show up
much in practice, however.

Change-Id: Ife1f018bac3c813f866e09dddec00cf8400c1083
Reviewed-on: http://review.northscale.com/1165
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>