ns_memcached improvements
Made it so ns_memcached is now always connected as long as the
gen_server is running, and we now continually retry until we're
connected at initial startup rather than crashing, which reduces
ugliness in the logs. The process doesn't register or start the
gen_server until it successfully connects. We still crash on errors
after the initial connection.

Also added wait_for_connection/1 to wait for memcached to successfully
come up on a specified set of nodes, to prevent the problem where
memcached is still coming up when the user hits rebalance.

Change-Id: Ic06a83776fb9bc08e75a61189a0b223a914c7c08
Reviewed-on: http://review.northscale.com/1437
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>