Refactor ns_memcached to run per bucket and manage bucket creation.
This seems to be the best way to manage which buckets are active on
nodes. ns_memcached will not register until warmup is complete, but it
avoids blocking a thread with warmup by setting wait_for_warmup=false.

We also now verify vbucket states before attempting to move a vbucket,
in order to avoid accidentally destroying the user's data. We'll only
move from an active location to a non-active location, making sure to
delete the vbucket first.

Still need to actually change the size if we see it change.

Change-Id: Ia9209fa78d4616f732da4d140a56e75bee33a55b
Reviewed-on: http://review.northscale.com/1900
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>