Bug 2126: Properly set vbucket state during failover.
They're all set in parallel since we attempt to set the vbucket on the
failing node to dead and this can take a while.

This also tries to avoid errors' being seen by clients by setting the
vbucket on the new master to pending before setting it to dead on the
old master.

Change-Id: I5d0bbc4e759245e799d8f508f5dadcd38281523e
Reviewed-on: http://review.northscale.com/2010
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>