Bug 2113: Select the bucket after creating it.
When a bucket didn't already exist in memcached, ns_bucket would
create it and then fail to select it, meaning the next command would
cause a disconnect. If a rebalance or other operation was waiting,
this would allow it to get past the starting gate, potentially sending
a bunch of messages that would never get responses.

Change-Id: I6bbf37dc0cf78af1f110e994c5e59d1c8814e4d8
Reviewed-on: http://review.northscale.com/2032
Reviewed-by: Sean Lynch <seanl@literati.org>
Tested-by: Sean Lynch <seanl@literati.org>