Bug 2128: Use a local_content Mnesia table for stats.
This is actually much cleaner and results in less network traffic and
greater stability; I should have done it this way all along. It also
eliminates the need to rename tables during node rename, which wasn't
working correctly the way I was doing it. I'll eliminate the broken
table-renaming code in the next commit.

Change-Id: I2a61ce1a6658b45ccd3a9747e600d61ee39b93e4
Reviewed-on: http://review.northscale.com/2074
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>