Don't take down the whole server if ns_cluster crashes
I took it out of the supervision tree, instead creating a separate
supervisor with a way to turn ns_server_sup off and on. Ideally, we
shouldn't have anything but a supervisor supervising any critical part
of the app.

In addition, since supervisor:terminate_child and
supervisor:restart_child are both synchronous and we don't need to
listen for the exit message from the child, ns_cluster no longer needs
to be a gen_fsm, so I converted it to a gen_server.

Change-Id: Id70bab48992220f6375e85278a229e305a669985
Reviewed-on: http://review.northscale.com/2496
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>