Convert ns_mnesia to a gen_server.
This removes responsibility for making sure mnesia is running or
stopped at the right time from ns_cluster, instead handling it through
the normal supervision tree route.

To cluster mnesia together on the nodes, instead of adding the new
node to the mnesia config at join time, the new node notices it has no
schema and connects to the rest of the cluster on its own if there are
connected nodes.

Change-Id: I8f20f4fdf383c923ae55a51905cb2354bb9c6803
Reviewed-on: http://review.northscale.com/2542
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>