Set max_wait_for_decision to 10 seconds
This will prevent mnesia from hanging forever (at least on startup) if
a node goes down while a transaction is in an unclear state.

Right now the only table that we do heavyweight transactions on and
that is replicated to multiple nodes is the schema, so this could only
happen if a node went down while we were creating tables. However, I
plan to have more replicated tables in the future, such as configs and
logs.

The tradeoff is that we'll get an inconsistent_database event, at
which point we'll just need to pick a node to load from.

Change-Id: I5a1260903ba4652a5d15a3f1862d7e67f95e31ba
Reviewed-on: http://review.northscale.com/2543
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>