Bug 2390/2382: track memory and evictions for memcached buckets
I renamed the ep_num_value_ejects stat to evictions (text is the same)
and added a function to rename bytes to mem_used (memcached) and
ep_num_value_ejects to evictions (membase). I updated the javascript
to use evictions.

The text still needs to be updated for memcached buckets, and the
disk-related stuff removed.

Change-Id: I12defb6aa2c2ca47b24fe0120baf1a40192cc832
Reviewed-on: http://review.northscale.com/2563
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>