Bug 966: log a bunch of stuff at startup.
I log the os type and version and the output of ns_info:runtime,
though ns_info:runtime contains all the same information in a
different form.

Also removed system_info(procs) from the output since it's formatted
in an unreadable fashion. This would probably be good to have in a
reasonably formatted fashion, however.

Change-Id: I385e5641281753bcee18fd4e5260e70e6bb1b0c4
Reviewed-on: http://review.northscale.com/2604
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>