Consistency with node/server terminology.  Bug 2094.
This corrects several dialogs which are behind a button labeled
server to continue to use the same term server.

There is also a northscale->membase change and some changes to error
messages.  The term node is still sometimes used from the REST
error handling as there it is not expected to be as confusing and
frequently part of the parameters includes a "node".

Change-Id: I8a4c9f501732690761b681e24052bdac9362968c
Reviewed-on: http://review.northscale.com/2769
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>