Require username and password whenever changing settings. Bug 2353.
This change actually fell out of an edge case of bug 2353.  We used
to allow no admin credentials into the cluster, so we also had a way
to un-set them.  However, we also changed it so the cluster considered
itself unconfigured if the username and password weren't set.  In
testing REST for changing the port, the cluster was accidentally
unconfigured.

This change will enforce the setting of a username and password.

Change-Id: I885c238d19a30f35a277d4a618e7094cbf56b8b8
Reviewed-on: http://review.northscale.com/2799
Tested-by: Matt Ingenthron <matt@northscale.com>
Reviewed-by: Matt Ingenthron <matt@northscale.com>