Bug 2662: Don't barf if there are already replica vbuckets
I'd left out a function clause for the case where there was already a
replica by making the existing clause not match the replica
state. Instead of adding another clause, I just filter replicas out of
the list.

This also eliminates the check for the node being in
nodes_actual_proper, thus eliminating a call to the config server per
node we're replicating from every 10 seconds on the node running the
orchestrator. Instead, the caller is responsible for checking that the
called node is alive if it cares.

Change-Id: Ie38d4af41edb3e542c32e6cb649b61a53af8e656
Reviewed-on: http://review.northscale.com/3045
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>