collect node's /proc/meminfo if it's available
Once per minute we'll fetch /proc/meminfo if it's available and send
it to ns_doctor. That information is periodically logged and is
available in /diag. /proc/meminfo contains useful high-level overview
of memory situation on the system.

This also makes ns_heart:expensive_checks/0 exported, so that it's
accessible from erlang shell.

Change-Id: Iff844fd42e696409c89726684307ee05641c808f
Reviewed-on: http://review.northscale.com/3167
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Sean Lynch <seanl@literati.org>