Bug 2736: Set servers to correct list of nodes before rebalancing
I've actually changed the order of operations so we only loop through
the list of buckets once. This prevents us from loading all the
buckets at once right at the beginning, which causes poor performance
when there are a lot of buckets. The tradeoff is that now instead of a
pause right at the beginning you'll see a pause every nth of the way
through, where n is the number of buckets.

Change-Id: Ida6877b13d07298256e7a782d9f76f0e1e16abc4
Reviewed-on: http://review.northscale.com/3178
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>