Don't delete destination vbuckets before transferring them
Deleting them and then dumping a lot of data into them can cause a
memory overrun because persistence is currently blocked during
deletion.

This will result in deleted data's possibly being resurrected if the
deletes occur during rebalance and a transfer is retried, or just
before rebalance and a vbucket master is moved on top of a previous
replica.

Change-Id: I4353b37a02b2fba090c1485d241ef3bd6e08de10
Reviewed-on: http://review.membase.org/3352
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>