made CallbackSlot#broadcast reentrant
The approach is to allow destructive updates of 'slaves' property only
when broadcasting counter is 0. We don't reset slaves array to empty
list at the beginning of broadcast, but instead we do extra pass over
it at the end of broadcast (only if broadcasting became zero) to clean
dead slaves.

Change-Id: I401991ddfb2956846ebc20140e05b9bb41b102a6
Reviewed-on: http://review.membase.org/3479
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>