Always create replicas, even if rebalance fails.
This will ensure that at least vbuckets that were successfully moved
will have replicas. This won't help if the replicas are situated on a
node that started having problems during rebalance, but if the masters
are located on a node that's also having trouble, there's not much we
could do to fix that.

Change-Id: I7eb747fd2308c46d434aa43413eca759561af3fd
Reviewed-on: http://review.membase.org/3578
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>