MB-3119 - avoid blacklisting a restored server
moxi's configured with downstream_conn_max of 16.  When moxi's holding
onto 16 stale downstream conns for a "host:port:bucket" in its conn
pool, and that node is restored (added back after a removal), the 17th
attempt to either 'reuse a conn from the conn pool or create a new
conn' will succeed, avoiding a blacklisting.

Change-Id: Ia77445fd17cecc54563d74f673fb470756cf6f6f
Reviewed-on: http://review.membase.org/4001
Tested-by: Sean Lynch <seanl@literati.org>
Reviewed-by: Sean Lynch <seanl@literati.org>