MB-3160: sync config before leave
This will hopefully prevent the case where a node shows up as active
after being failed, rebalanced out, and readded, but I was not able to
reproduce the problem so I have no way to test that it actually
works. It will, however, prevent lots of potential problems when a
node leaves the cluster right after making some other config change,
so we should apply it.

Change-Id: Id251fdeaf99dbb7d235dcda9302864db173bb77a
Reviewed-on: http://review.membase.org/4055
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>