MB-3195: Shut down replication during rebalance
Only shuts down the bucket we're rebalancing, and not until rebalance
actually starts for that bucket.

Should improve performance, reduce failures, and eliminate the
spurious vbucketmigrator error messages in the log.

This creates a potential window of dataloss but leaving it running
without intelligent decisions about where to move data is even
riskier.

Change-Id: I48aac7542563862eceee9877aef7544e171f53d6
Reviewed-on: http://review.membase.org/4098
Tested-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>