Automatic clustering and declustering of Mnesia.
Add ns_mnesia:promote/1, promote_self/1, and demote/1. promote/ and
demote/1 must be called from a node that's already clustered and isn't
the node being manipulated. promote_self/1 must be called from the new
node.

Both promotion and demotion require that Mnesia be restarted on the
node that is changing state, but not on any existing clustered
nodes. At the moment, the worst this causes is the dropping of a few
stats samples.

Change-Id: I36de5bb4227ede60dc9100f428d4413168baa72c
Reviewed-on: http://review.membase.org/4432
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>