Leader election.
This starts mb_master_sup on the lowest-numbered candidate node that's
alive, broadcasting which node is the master to all nodes listed in
nodes_wanted. Processes wanting to talk to processes that run under
mb_master_sup should use mb_master:master_node() to determine which
node is the master, then send messages to a registered name on that
node.

Change-Id: Ib75bffd03bf2526463c216056cf166e0cc7dc859
Reviewed-on: http://review.membase.org/4631
Reviewed-by: Sean Lynch <seanlynch5@gmail.com>
Tested-by: Sean Lynch <seanlynch5@gmail.com>