Define timeout for vbucket deletion command.
We have a few cases where the timeout is actually occuring in a
production environment where the command would otherwise succeed, but
does so too slowly here.  This code adjusts the timeout up to a larger
value so we might complete instead of just timing out.

The code that calls this actually specifies a timeout (~30s), but it's
not included down here.

Timeout (as seen in a production log):

    ** Generic server <5596.169.0> terminating
    ** Last message in was {delete_vbucket,857}
    ** When Server state == {state,"default",#Port<5596.2633>}
    ** Reason for termination ==
    ** {{badmatch,{error,timeout}},
        [{mc_client_binary,cmd_binary_vocal_recv,5},
         {mc_client_binary,delete_vbucket,2},
         {ns_memcached,handle_call,3},
         {gen_server,handle_msg,5},
         {proc_lib,init_p_do_apply,3}]}

Change-Id: I945a4670e0f5ab135f710a083068d7953e8ac137